requirements.txt
streamlit
Pillow
pytesseract
requests
beautifulsoup4
tensorflow